Improve logic for powersaving and prevent double runs
these double runs where happing because udev is soo fast that it
literally takes microseconds to execute rules not giving battery adaptor
enough time to transition from Discharging -> Unknown -> Charging.

So usually during the time of executing rule state is at Unknown which
isn't triggered by rule any of the rules leading to system still in
previous state which is usually battery mode.

I could have gone the path of adding Unknown but Unknown is median way
through Discharging and Charging so if used will be triggered every time.

similarly due to the issue of udev been too fast,Charging is always executed when power
source is taking out before state changes to Unknown leading to both Charging and no power source
rules executing but the latter taking effect

so now we use the more relaiable ADP[0-9] adaptor for both detecting when
power supply is connected and or disconnected which is its main function

we guard the power supply ON/OFF rules to ensure that they are executed
only when the required conditions are meet