sway(launcher): Fix the command in the enter:execute bind
The issue was that when the $ was not escaped in the double quotes
It was treated as a variable reference which lead to immediate expansion
resulting in an empty state,Because at the time of execution the user
hasn't yet given an input.

To prevent this immediate execution, we escape the $ symbol so it is
treated literally when parsing, resulting in the execution of the
command as a command .ie $(<$user_input) and at the right time when
$user_input has been populated. Also referencing state we muse escape
the $ for the same reason so that it is expanded only after the
execution of the read extraction command

Do these for every where you reference a $ symbol which is a special
symbol in a double quote.